% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/r_next_srd.R
\name{cr_next_srd}
\alias{cr_next_srd}
\alias{r_next_srd}
\title{Generate Next Sample of Square-Root Diffusion}
\usage{
cr_next_srd(x_0, k, theta, sigma, h)

r_next_srd(x_0, k, theta, sigma, h = 1)
}
\arguments{
\item{x_0}{current value of the Square-Root Diffusion.}

\item{k}{parameter k, mean-reverting velocity parameter.}

\item{theta}{parameter \eqn{\theta}, long-run mean.}

\item{sigma}{parameter \eqn{\sigma}, volatility parameter.}

\item{h}{time interval length.}
}
\value{
Next sample of the Square-Root Diffusion, a scale value.
}
\description{
Generate next sample of Square-Root Diffusion given a current value x_0.
}
\details{
Square-Root Diffusion is described by the following stochastic differential
equation:
\deqn{dx(t) = k(\theta - x(t))dt + \sigma\sqrt(x(t))dw(t),}
where \eqn{k,\theta,\sigma} are parameters with positive values and
satisfy the condition \eqn{\sigma^2 \leq 2k\theta}, \eqn{w(t)} is a
standard Wiener process. In addition, the initial value of \eqn{x(t)}
should be positive, i.e., \eqn{x(0) > 0}.

Given \eqn{x(s)}, the transition distribution of
\eqn{x(t)} for \eqn{t > s} is a noncentral chi-squared distribution
multiplied with a scale factor (see Cox et al. 1985), i.e.,
\deqn{x(t) = \frac{\sigma^2[1-e^{-k(t-s)}]}{4k} \chi'^2_d
      \left(\frac{4k e^{-k(t-s)}}
      {\sigma^2[1-e^{-k(t-s)}]}x(s) \right)
      ,\quad s < t, }
where \eqn{\chi'^2_d(\lambda)} represents the noncentral chi-squared
random variable with \eqn{d} degrees of freedom and non-centrality
parameter \eqn{\lambda}, and \eqn{d=4k_i\theta_i/\sigma_{xi}^2}.
}
\examples{
k = 2; theta = 1.5; sigma = 1.6; v_0 = theta; h = 1
r_next_srd(v_0, k, theta, sigma, h)
}
