% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/r_srd.R
\name{cr_srd}
\alias{cr_srd}
\alias{r_srd}
\title{Generate Samples of Square-Root Diffusion}
\usage{
cr_srd(x_0, k, theta, sigma, N, h)

r_srd(x_0, k, theta, sigma, N, h = 1)
}
\arguments{
\item{x_0}{current value of the Square-Root Diffusion.}

\item{k}{parameter k, mean-reverting velocity parameter.}

\item{theta}{parameter \eqn{\theta}, long-run mean.}

\item{sigma}{parameter \eqn{\sigma}, volatility parameter.}

\item{N}{length of samples to generate.}

\item{h}{time interval length.}
}
\value{
A series of N samples of the Square-Root Diffusion, a vector.
}
\description{
Generate samples of Square-Root Diffusion.
}
\details{
Square-Root Diffusion is described by the following stochastic differential
equation:
\deqn{dx(t) = k(\theta - x(t))dt + \sigma\sqrt{x(t)}dw(t),}
where \eqn{k,\theta,\sigma} are parameters with positive values and
satisfy the condition \eqn{\sigma^2 \leq 2k\theta}, \eqn{w(t)} is a
standard Wiener process. In addition, the initial value of \eqn{x(t)}
should be positive, i.e., \eqn{x(0) > 0}.

Given \eqn{x(s)}, the transition distribution of
\eqn{x(t)} for \eqn{t > s} is a noncentral chi-squared distribution
multiplied with a scale factor (see Cox et al. 1985), i.e.,
\deqn{x(t) = \frac{\sigma^2[1-e^{-k(t-s)}]}{4k} \chi'^2_d
      \left(\frac{4k e^{-k(t-s)}}
      {\sigma^2[1-e^{-k(t-s)}]}x(s) \right)
      ,\quad s < t, }
where \eqn{\chi'^2_d(\lambda)} represents the noncentral chi-squared
random variable with \eqn{d} degrees of freedom and non-centrality
parameter \eqn{\lambda}, and \eqn{d=4k\theta/\sigma^2}.
}
\examples{
k = 2; theta = 1.5; sigma = 1.6
v_0 = theta; N = 200; h = 1
r_srd(v_0, k, theta, sigma, N, h)
}
\seealso{
Sample one-step transition refers to \link[=r_next_srd]{r_next_srd}.
}
